+++ jest.config.base.js
@@ -4,7 +4,11 @@ module.exports = {
   roots: ["<rootDir>/src"],
   transform: {
     "^.+\\.ts$": "ts-jest",
+    "node_modules/chai/.+\\.js$": "ts-jest",
   },
+  transformIgnorePatterns: [
+    "/node_modules/(?!(chai)/)",
+  ],
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
   collectCoverage: true,
   verbose: true,