+++ packages/xrpl/src/models/index.ts
@@ -13,6 +13,11 @@ export {
   convertTxFlagsToNumber,
   parseTransactionFlags,
 } from './utils/flags'
+export {
+  validateMPTokenMetadata,
+  decodeMPTokenMetadata,
+  encodeMPTokenMetadata,
+} from './utils/mptokenMetadata'
 export * from './methods'
 export * from './transactions'
 export * from './common'